requirements.txt
selenium
pytest-bdd
pytest-splinter
